+++ ati_netcanoem_ft_driver/CMakeLists.txt.ros2
@@ -17,7 +17,7 @@ set(Eigen3_INCLUDE_DIRS ${EIGEN3_INCLUDE_DIR})
 
 ## Specify additional locations of header files
 ## Your package locations should be listed before other locations
-include_directories(include SYSTEM ${Eigen3_INCLUDE_DIRS})
+include_directories(include src SYSTEM ${Eigen3_INCLUDE_DIRS})
 
 ## Build options
 add_compile_options(-std=c++14)
@@ -39,18 +39,22 @@ add_library(${PROJECT_NAME}
             src/${PROJECT_NAME}/ati_netcanoem_ft_driver.cpp)
 ament_target_dependencies(${PROJECT_NAME} common_robotics_utilities)
 
-## Declare C++ components
-add_library(${PROJECT_NAME}_components SHARED src/ati_netcanoem_ft_driver_node.ros2.cpp)
+## Declare C++ components library
+add_library(${PROJECT_NAME}_components SHARED
+            src/ati_netcanoem_ft_driver_node.ros2.cpp)
 target_link_libraries(${PROJECT_NAME}_components ${PROJECT_NAME})
 ament_target_dependencies(${PROJECT_NAME}_components
   geometry_msgs std_srvs rclcpp rclcpp_components)
 
 ## Register C++ components
-rclcpp_components_register_node(
+rclcpp_components_register_nodes(
   ${PROJECT_NAME}_components
-  PLUGIN "ati_netcanoem_ft_driver::AtiNetCanOemDriverNode"
-  EXECUTABLE ${PROJECT_NAME}_node
+  "ati_netcanoem_ft_driver::AtiNetCanOemDriverNode"
 )
+
+## Declare C++ component executable
+add_executable(${PROJECT_NAME}_node src/ati_netcanoem_ft_driver_node_main.ros2.cpp)
+target_link_libraries(${PROJECT_NAME}_node ${PROJECT_NAME}_components)
 
 ## Declare a C++ executable
 add_executable(configure_ati_can_ft_sensor src/configure_ati_can_ft_sensor.cpp)
@@ -68,6 +72,11 @@ install(TARGETS
   ARCHIVE DESTINATION lib
   LIBRARY DESTINATION lib
   RUNTIME DESTINATION bin
+)
+
+install(
+  TARGETS ${PROJECT_NAME}_node
+  RUNTIME DESTINATION lib/${PROJECT_NAME}
 )
 
 ## Mark cpp header files for installation
